Time: 2800 ms (5.06%), Space: 422.6 MB (5.01%) - LeetHub

+++ 115-distinct-subsequences/115-distinct-subsequences.cpp
@@ -51,7 +51,7 @@ public:
             }
             prev = curr;
         }
-        return prev[m];
+        return (int)prev[m];
         
     }
 };
